Made a number of changes to feed display and entry subscription
Allowed the user to save entries from the featured page to their own feed. This feed can then be browsed in the "My Feeds" area. Also added functionality for adding articles individually by URL, though this is not yet available in the UI. Subscriptions have been added in the database to allow a user to build a feed where subscriptions are made to other pre-built feeds, though this is also not yet available in the UI. A new database script is included.